requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.5.0
requests>=2.31.0
transformers>=4.30.0
accelerate>=0.20.0
sounddevice>=0.4.6
soundfile>=0.12.1
pydub>=0.25.1
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pytesseract>=0.3.10
python-dotenv>=1.0.0
pyaudio>=0.2.13